Give site-page-items its own test file, and stop the nav tests repeating
The mutation gate wants every changed source to have a test at the
matching path. site_page_items had none: its tests lived inside the
site_pages file, which had grown to 383 lines. Moving them gives each
source its mirror and leaves two files well under the limit.

The moved tests only ever needed "a page exists", so they use the shared
createTestSitePage helper rather than carrying a second page-making
helper across.

The two site-nav leaf tests shared their whole setup, which jscpd caught.
One helper places a listing on a page and reads the nav back, taking an
optional change to make to the listing first.
